FW-2596: move panic catcher code into its own panic module
